Finish refactor:move topicQuery SQL database interrogation to articles.models
