+++ app/controllers/articles.controllers.js
@@ -1,5 +1,3 @@
-const db = require("../../db/connection");
-
 const {
   selectArticleByArticleId,
   selectAllArticles,
@@ -8,6 +6,7 @@ const {
   insertCommentToArticle,
   insertNewArticle,
 } = require("../models/articles.models");
+const { selectTopicBySlug } = require("../models/topics.models");
 
 const getArticleByArticleId = (req, res, next) => {
   const { article_id } = req.params;
@@ -20,82 +19,24 @@ const getArticleByArticleId = (req, res, next) => {
 
 const getAllArticles = (req, res, next) => {
   const { sort_by, order, topic } = req.query;
-  //function to check topic existence - rejects if it does not
-  const checkTopicExists = (topicName) => {
-    return db
-      .query(`SELECT * FROM topics WHERE slug = $1`, [topicName])
-      .then(({ rows }) => {
-        if (rows.length === 0) {
-          return Promise.reject({ status: 404, msg: "Not found!" });
-        } else {
-          return true;
-        }
+
+  if (topic) {
+    const pendingSelectTopicBySlug = selectTopicBySlug(topic);
+    const pendingSelectAllArticles = selectAllArticles(sort_by, order, topic);
+    return Promise.all([pendingSelectAllArticles, pendingSelectTopicBySlug])
+      .then((articles) => {
+        res.status(200).send({ articles: articles[0] });
       })
       .catch(next);
-  };
-
-  if (topic) {
-    checkTopicExists(topic);
-  }
-  let queryStr = `WITH mainCommentsInfo AS
-          (SELECT author, title, article_id, topic, created_at, votes, article_img_url
-          FROM articles),
-
-          commentCountInfo AS
-          (SELECT article_id, COUNT(comment_id)::INT
-          AS comment_count
-          FROM comments
-          GROUP BY article_id)
-
-          SELECT mainCommentsInfo.author,
-          mainCommentsInfo.title, mainCommentsInfo.article_id, mainCommentsInfo.topic, mainCommentsInfo.created_at, mainCommentsInfo.votes, mainCommentsInfo.article_img_url,
-          COALESCE(commentCountInfo.comment_count, 0) AS comment_count
-          FROM mainCommentsInfo
-
-          LEFT JOIN commentCountInfo 
-          ON mainCommentsInfo.article_id = commentCountInfo.article_id`;
-  const sortByGreenlist = ["created_at", "votes", "topic", "author", "title"];
-  const orderGreenlist = ["asc", "desc"];
-  const sortByRedlist = ["body", "article_img_url"];
-  //topic handling
-  if (topic) {
-    queryStr += ` WHERE topic = '${topic}'`;
-  }
-  //sort_by handling
-  if (!sort_by) {
-    queryStr += ` ORDER BY created_at`;
-  }
-  if (sort_by) {
-    if (sortByGreenlist.includes(sort_by)) {
-      queryStr += ` ORDER BY ${sort_by}`;
-    }
-    if (sortByRedlist.includes(sort_by)) {
-      return Promise.reject({ status: 401, msg: "Unauthorised request!" });
-    }
-  }
-  //400 rejects
-  if (
-    (sort_by && !sortByGreenlist.includes(sort_by)) ||
-    sort_by === "" ||
-    (order && !orderGreenlist.includes(order.toLowerCase())) ||
-    order === "" ||
-    topic === ""
-  ) {
-    return Promise.reject({ status: 400, msg: "Invalid request!" });
   }
 
-  //order handling
-  if (!order) {
-    queryStr += ` DESC`;
+  if (!topic) {
+    return selectAllArticles(sort_by, order, topic)
+      .then((articles) => {
+        res.status(200).send({ articles });
+      })
+      .catch(next);
   }
-  if (order && orderGreenlist.includes(order.toLowerCase())) {
-    queryStr += ` ${order}`;
-  }
-  return selectAllArticles(queryStr)
-    .then((articles) => {
-      res.status(200).send({ articles });
-    })
-    .catch(next);
 };
 
 const patchVotesOnArticle = (req, res, next) => {
