+++ app/models/articles.models.js
@@ -29,7 +29,61 @@ const selectArticleByArticleId = (articleId) => {
     });
 };
 
-const selectAllArticles = (queryStr) => {
+const selectAllArticles = (sort_by, order, topic) => {
+  let queryStr = `WITH mainCommentsInfo AS
+  (SELECT author, title, article_id, topic, created_at, votes, article_img_url
+  FROM articles),
+
+  commentCountInfo AS
+  (SELECT article_id, COUNT(comment_id)::INT
+  AS comment_count
+  FROM comments
+  GROUP BY article_id)
+
+  SELECT mainCommentsInfo.author,
+  mainCommentsInfo.title, mainCommentsInfo.article_id, mainCommentsInfo.topic, mainCommentsInfo.created_at, mainCommentsInfo.votes, mainCommentsInfo.article_img_url,
+  COALESCE(commentCountInfo.comment_count, 0) AS comment_count
+  FROM mainCommentsInfo
+
+  LEFT JOIN commentCountInfo 
+  ON mainCommentsInfo.article_id = commentCountInfo.article_id`;
+  const sortByGreenlist = ["created_at", "votes", "topic", "author", "title"];
+  const orderGreenlist = ["asc", "desc"];
+  const sortByRedlist = ["body", "article_img_url"];
+  //topic handling
+  if (topic) {
+    queryStr += ` WHERE topic = '${topic}'`;
+  }
+  //sort_by handling
+  if (!sort_by) {
+    queryStr += ` ORDER BY created_at`;
+  }
+  if (sort_by) {
+    if (sortByGreenlist.includes(sort_by)) {
+      queryStr += ` ORDER BY ${sort_by}`;
+    }
+    if (sortByRedlist.includes(sort_by)) {
+      return Promise.reject({ status: 401, msg: "Unauthorised request!" });
+    }
+  }
+  //400 rejects
+  if (
+    (sort_by && !sortByGreenlist.includes(sort_by)) ||
+    sort_by === "" ||
+    (order && !orderGreenlist.includes(order.toLowerCase())) ||
+    order === "" ||
+    topic === ""
+  ) {
+    return Promise.reject({ status: 400, msg: "Invalid request!" });
+  }
+  //order handling
+  if (!order) {
+    queryStr += ` DESC`;
+  }
+  if (order && orderGreenlist.includes(order.toLowerCase())) {
+    queryStr += ` ${order}`;
+  }
+
   return db.query(queryStr).then(({ rows: articles }) => {
     return articles;
   });
